+++ tests/_test.py
@@ -29,7 +29,7 @@ from . import (
 
 def test_color_string():
     """Test a simple string."""
-    color = Color(text="green")
+    color = Color(fore="green")
     assert color.get(TEST_STR) == f"{GREEN}{TEST_STR}{RESET}"
 
 
@@ -51,34 +51,34 @@ def test__dir__(populated_colors):
     color_instances = populated_colors.__dir__()
     assert color_instances == INSTANCES
     color = Color("red", "bold", "green")
-    assert color.text == 1
+    assert color.fore == 1
     assert color.effect == 1
-    assert color.background == 2
+    assert color.back == 2
 
 
 def test_str_args():
     """Test for the correct ANSI codes for color, effect, and
-    background
+    background.
     """
     _color = Color("red", "bold", "green")
-    assert _color.text == 1
+    assert _color.fore == 1
     assert _color.effect == 1
-    assert _color.background == 2
+    assert _color.back == 2
 
 
 def test_str_ints():
     """Test the correct ANSI code integers are produced."""
     _color = Color(1, 1, 2)
-    assert _color.text == 1
+    assert _color.fore == 1
     assert _color.effect == 1
-    assert _color.background == 2
+    assert _color.back == 2
 
 
 def test_int_dict():
     """Test for a non-existing color and assert it will default to
     white instead of raising an error."""
     _color = Color(orange=1)
-    assert _color.text == 7
+    assert _color.fore == 7
 
 
 def test_exact_letter_in_colored_string(populated_colors):
@@ -237,7 +237,7 @@ def test_none():
     """Test the correct string is returned when no keyword is used for
     multicolor.
     """
-    _color = Color(text="GREEN")
+    _color = Color(fore="GREEN")
     none = _color.multicolor(LONG_STRING)
     assert none == LONG_STRING